Move the `monitor` function to the domain layer: it only imports domain modules from our project now.
