+++ walkingpadfitbit/domain/monitor.py
